Extract common `agent` subcommands into packages
This follows the same steps of #13967, but for `config`, `health`, and
`version`, which are all used by the DCA as well.
