+++ pkg/cli/subcommands/config/command_test.go
@@ -8,16 +8,22 @@ package config
 import (
 	"testing"
 
+	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/require"
 
-	"github.com/DataDog/datadog-agent/cmd/agent/command"
 	"github.com/DataDog/datadog-agent/comp/core"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
 func TestConfigCommand(t *testing.T) {
+	commands := []*cobra.Command{
+		MakeCommand(func() GlobalParams {
+			return GlobalParams{}
+		}),
+	}
+
 	fxutil.TestOneShotSubcommand(t,
-		Commands(&command.GlobalParams{}),
+		commands,
 		[]string{"config"},
 		showRuntimeConfiguration,
 		func(cliParams *cliParams, coreParams core.BundleParams) {
@@ -27,8 +33,14 @@ func TestConfigCommand(t *testing.T) {
 }
 
 func TestConfigListRuntimeCommand(t *testing.T) {
+	commands := []*cobra.Command{
+		MakeCommand(func() GlobalParams {
+			return GlobalParams{}
+		}),
+	}
+
 	fxutil.TestOneShotSubcommand(t,
-		Commands(&command.GlobalParams{}),
+		commands,
 		[]string{"config", "list-runtime"},
 		listRuntimeConfigurableValue,
 		func(cliParams *cliParams, coreParams core.BundleParams) {
@@ -38,8 +50,14 @@ func TestConfigListRuntimeCommand(t *testing.T) {
 }
 
 func TestConfigSetCommand(t *testing.T) {
+	commands := []*cobra.Command{
+		MakeCommand(func() GlobalParams {
+			return GlobalParams{}
+		}),
+	}
+
 	fxutil.TestOneShotSubcommand(t,
-		Commands(&command.GlobalParams{}),
+		commands,
 		[]string{"config", "set", "foo", "bar"},
 		setConfigValue,
 		func(cliParams *cliParams, coreParams core.BundleParams) {
@@ -49,8 +67,14 @@ func TestConfigSetCommand(t *testing.T) {
 }
 
 func TestConfigGetCommand(t *testing.T) {
+	commands := []*cobra.Command{
+		MakeCommand(func() GlobalParams {
+			return GlobalParams{}
+		}),
+	}
+
 	fxutil.TestOneShotSubcommand(t,
-		Commands(&command.GlobalParams{}),
+		commands,
 		[]string{"config", "get", "foo"},
 		getConfigValue,
 		func(cliParams *cliParams, coreParams core.BundleParams) {
